Refactor CounterThres
Change the way players' throw stats are tracked:
Now, there's an array whose size is the number of all possible Throw
values. For each Throw tracked, the corresponding element of the array
is incremented.

Introduced behavior-controlling keyword arguments:
Minimum number of data points and throw frequency threshold can now be
set upon initialization of the CounterThresPlayer class.

Introduce the condition that exactly one most frequent Throw exist for a
player to be suspected to be a ThresPlayer.

Also extended unittest coverage.